+++ example-functional/src/main/java/com/dan323/functional/data/parser/ParserApplicative.java
@@ -13,7 +13,7 @@ import java.util.function.Supplier;
 
 @Applicative
 @Alternative
-public class ParserApplicative implements IApplicative<Parser<?>>, IAlternative<Parser<?>> {
+public final class ParserApplicative implements IApplicative<Parser<?>>, IAlternative<Parser<?>> {
 
     private static final StateMonad<String, FiniteList<Parser.ParserError>> STATE_MONAD = StateMonad.getInstance();
 
@@ -21,6 +21,9 @@ public class ParserApplicative implements IApplicative<Parser<?>>, IAlternative<
 
     public static ParserApplicative getInstance() {
         return PARSER_MONAD;
+    }
+
+    private ParserApplicative() {
     }
 
     @Override